backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0

# AI / ML
transformers==4.41.2
torch==2.3.0
Pillow==10.3.0

# Push notifications
pywebpush==2.0.0
py-vapid==1.9.0

# Scheduling
apscheduler==3.10.4

# Auth (JWT validation)
python-jose[cryptography]==3.3.0

# Rate limiting
slowapi==0.1.9

# Logging
loguru==0.7.2

# Dev / test
python-dotenv==1.0.1
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
ruff==0.4.4
